feat(transaction): adding feature deposit and withdraw for user

+++ wallet-service/internal/dto/wallet.go
@@ -12,13 +12,13 @@ type GetWalletRequest struct {
 
 type UpdateBalanceRequest struct {
 	UserID string  `json:"user_id"`
-	Amount float64 `json:"amount"`
+	Amount int64 `json:"amount"`
 	Operation string `json:"operation"`
 }
 
 type WalletResponse struct {
 	UserID string  `json:"user_id"`
-	Balance float64 `json:"balance"`
+	Balance int64 `json:"balance"`
 	Message string `json:"message"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
